resolving concurrency problems and validations

+++ src/main/java/com/felipe/fileuploader/daos/GenericDaoImpl.java
@@ -18,6 +18,7 @@ import javax.ws.rs.InternalServerErrorException;
 
 import com.felipe.fileuploader.entities.Entity;
 import com.felipe.fileuploader.util.AppConfiguration;
+import com.felipe.fileuploader.util.AppendableObjectOutputStream;
 import com.felipe.fileuploader.util.DirUtil;
 
 public class GenericDaoImpl<T extends Entity<I>, I extends Serializable>
@@ -31,24 +32,58 @@ public class GenericDaoImpl<T extends Entity<I>, I extends Serializable>
 
 	@Override
 	public Boolean save(T entity) {
-		
+
 		Boolean success = true;
 		FileOutputStream fout = null;
 		ObjectOutputStream oos = null;
 
 		try {
-			fout = new FileOutputStream(pathBaseName, true);
-			oos = new ObjectOutputStream(fout);
+
+			//List<T> allEntities = findAll();
+
+			if (new File(getDataBasePath()).exists()) {
+				fout = new FileOutputStream(pathBaseName, true);
+				oos = new AppendableObjectOutputStream(fout);
+			} else {
+				fout = new FileOutputStream(pathBaseName);
+				oos = new ObjectOutputStream(fout);
+			}
+
+			//fout = new FileOutputStream(pathBaseName, true);
+			//oos = new AppendableObjectOutputStream(fout);
+
+			// fout = new FileOutputStream(pathBaseName);
+			// oos = new ObjectOutputStream(fout);
+
+			// for (T previousEntity : allEntities) {
+			// oos.writeObject(previousEntity);
+			// }
+
 			oos.writeObject(entity);
+			// oos.reset();
 		} catch (IOException e) {
 			throw new InternalServerErrorException(AppConfiguration.get(
 					"error.entity_not_saved", entity.getClass()));
 		} finally {
+			if (fout != null) {
+				try {
+					fout.flush();
+					fout.close();
+				} catch (IOException e) {
+					throw new InternalServerErrorException(
+							AppConfiguration
+									.get("error.internal_error_message"));
+				}
+			}
+
 			if (oos != null) {
 				try {
+					oos.flush();
 					oos.close();
 				} catch (IOException e) {
-					e.printStackTrace();
+					throw new InternalServerErrorException(
+							AppConfiguration
+									.get("error.internal_error_message"));
 				}
 			}
 		}
@@ -64,11 +99,13 @@ public class GenericDaoImpl<T extends Entity<I>, I extends Serializable>
 		List<T> entities = new LinkedList<>();
 
 		FileInputStream fis = null;
+		ObjectInputStream ois = null;
 
 		try {
+			// TODO felipegc abri um trem aqui
 			fis = new FileInputStream(pathBaseName);
+			ois = new ObjectInputStream(fis);
 			while (true) {
-				ObjectInputStream ois = new ObjectInputStream(fis);
 				entities.add((T) ois.readObject());
 			}
 		} catch (EOFException ignored) {
@@ -84,6 +121,14 @@ public class GenericDaoImpl<T extends Entity<I>, I extends Serializable>
 					e.printStackTrace();
 				}
 			}
+
+			if (ois != null) {
+				try {
+					ois.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			}
 		}
 
 		return entities;
@@ -94,15 +139,15 @@ public class GenericDaoImpl<T extends Entity<I>, I extends Serializable>
 		if (!new File(getDataBasePath()).exists()) {
 			return null;
 		}
-		
+
 		Map<I, T> infos = new HashMap<>();
 
 		FileInputStream fis = null;
 
 		try {
 			fis = new FileInputStream(pathBaseName);
+			ObjectInputStream ois = new ObjectInputStream(fis);
 			while (true) {
-				ObjectInputStream ois = new ObjectInputStream(fis);
 				T objRead = (T) ois.readObject();
 				infos.put(objRead.getId(), objRead);
 			}
